Rename module
Moved the module to the package root as `index.ts` + adjust import
